agentfs run: Restructure the command files for clarity
cmd/ changes:
- Create run.rs as common entry point that dispatches to platform impls
- Rename run.rs → run_linux.rs (Linux FUSE+namespace or ptrace)
- Rename run_nfs.rs → run_darwin.rs (Darwin NFS+sandbox-exec)
- Delete run_stub.rs (fallback now inline in run.rs)
- Simplify mod.rs (no more #[cfg] dance)

sandbox/ changes:
- Rename overlay.rs → linux.rs (it's the Linux sandbox, not just overlay)
- Rename ptrace.rs → linux_ptrace.rs (clarify it's Linux-specific)
- Rename sandbox_macos.rs → darwin.rs (remove redundant prefix)

The new structure makes it clear which files are platform-specific
and provides a single entry point for the run command.
