+++ cli/src/cmd/mod.rs
@@ -8,20 +8,6 @@ mod mount;
 #[path = "mount_stub.rs"]
 mod mount;
 
-// Run module selection:
-// - Linux x86_64: use overlay sandbox (run.rs)
-// - macOS: use NFS-based sandbox (run_nfs.rs)
-// - Other platforms: use stub (run_stub.rs)
-
-#[cfg(target_os = "linux")]
-mod run;
-
-#[cfg(target_os = "macos")]
-#[path = "run_nfs.rs"]
-mod run;
-
-#[cfg(not(any(all(target_os = "linux"), target_os = "macos")))]
-#[path = "run_stub.rs"]
 mod run;
 
 // Standalone NFS server command (Unix only)
